datasets and fields roundtrip with new structure
