+++ arcana/dataset/__init__.py
@@ -1,7 +1,7 @@
 from .item import Dataset, Field
 from .collection import DatasetCollection, FieldCollection
 from .spec import DatasetSpec, FieldSpec, BaseSpec
-from .base import BaseField, BaseDataset, BaseDatasetOrField
+from .base import BaseField, BaseDataset, BaseData
 from .match import DatasetMatch, FieldMatch, BaseMatch
 from .file_format import FileFormat, Converter, IdentityConverter
 from .bids import BidsDatasetMatch, BidsAssociatedDatasetMatch
