+++ test/unittests/repository/test_local.py
@@ -50,218 +50,6 @@ class DummyStudy(with_metaclass(StudyMetaClass, Study)):
         pass
 
 
-class TestLocalRepository(BaseTestCase):
-
-    STUDY_NAME = 'astudy'
-    SUMMARY_STUDY_NAME = 'asummary'
-    INPUT_DATASETS = {'source1': '1',
-                      'source2': '2',
-                      'source3': '3',
-                      'source4': '4'}
-
-    def test_repository_roundtrip(self):
-        study = DummyStudy(
-            self.STUDY_NAME, self.repository, runner=LinearRunner('a_dir'),
-            inputs=[DatasetMatch('source1', text_format, 'source1'),
-                    DatasetMatch('source2', text_format, 'source2'),
-                    DatasetMatch('source3', text_format, 'source3'),
-                    DatasetMatch('source4', text_format, 'source4')])
-        # TODO: Should test out other file formats as well.
-        source_files = [study.input(n)
-                        for n in ('source1', 'source2', 'source3',
-                                  'source4')]
-        sink_files = [study.spec(n)
-                      for n in ('sink1', 'sink3', 'sink4')]
-        inputnode = pe.Node(IdentityInterface(['subject_id', 'visit_id']),
-                            'inputnode')
-        inputnode.inputs.subject_id = self.SUBJECT
-        inputnode.inputs.visit_id = self.VISIT
-        source = self.repository.source(source_files,
-                                     study_name=self.STUDY_NAME)
-        sink = self.repository.sink(sink_files, study_name=self.STUDY_NAME)
-        sink.inputs.name = 'repository_sink'
-        sink.inputs.desc = (
-            "A test session created by repository roundtrip unittest")
-        # Create workflow connecting them together
-        workflow = pe.Workflow('source_sink_unit_test', base_dir=self.work_dir)
-        workflow.add_nodes((source, sink))
-        workflow.connect(inputnode, 'subject_id', source, 'subject_id')
-        workflow.connect(inputnode, 'visit_id', source, 'visit_id')
-        workflow.connect(inputnode, 'subject_id', sink, 'subject_id')
-        workflow.connect(inputnode, 'visit_id', sink, 'visit_id')
-        for source_file in source_files:
-            if not source_file.name.endswith('2'):
-                source_name = source_file.name
-                sink_name = source_name.replace('source', 'sink')
-                workflow.connect(
-                    source, source_name + PATH_SUFFIX,
-                    sink, sink_name + PATH_SUFFIX)
-        workflow.run()
-        # Check local directory was created properly
-        outputs = [
-            f for f in sorted(os.listdir(self.session_dir))
-            if f != FIELDS_FNAME]
-        self.assertEqual(outputs,
-                         [self.STUDY_NAME + '_sink1.txt',
-                          self.STUDY_NAME + '_sink3.txt',
-                          self.STUDY_NAME + '_sink4.txt',
-                          'source1.txt', 'source2.txt',
-                          'source3.txt', 'source4.txt'])
-
-    def test_fields_roundtrip(self):
-        repository = LocalRepository(base_dir=self.project_dir)
-        sink = repository.sink(
-            outputs=[
-                FieldSpec('field1', int, 'pipeline'),
-                FieldSpec('field2', float, 'pipeline'),
-                FieldSpec('field3', str, 'pipeline')],
-            name='fields_sink',
-            study_name='test')
-        sink.inputs.field1_field = field1 = 1
-        sink.inputs.field2_field = field2 = 2.0
-        sink.inputs.field3_field = field3 = '3'
-        sink.inputs.subject_id = self.SUBJECT
-        sink.inputs.visit_id = self.VISIT
-        sink.inputs.desc = "Test sink of fields"
-        sink.inputs.name = 'test_sink'
-        sink.run()
-        source = repository.source(
-            inputs=[
-                FieldSpec('field1', int, 'dummy_pipeline'),
-                FieldSpec('field2', float, 'dummy_pipeline'),
-                FieldSpec('field3', str, 'dummy_pipeline')],
-            name='fields_source',
-            study_name='test')
-        source.inputs.visit_id = self.VISIT
-        source.inputs.subject_id = self.SUBJECT
-        source.inputs.desc = "Test source of fields"
-        source.inputs.name = 'test_source'
-        results = source.run()
-        self.assertEqual(results.outputs.field1_field, field1)
-        self.assertEqual(results.outputs.field2_field, field2)
-        self.assertEqual(results.outputs.field3_field, field3)
-
-    def test_summary(self):
-        study = DummyStudy(
-            self.SUMMARY_STUDY_NAME, self.repository, LinearRunner('ad'),
-            inputs=[DatasetMatch('source1', text_format, 'source1'),
-                    DatasetMatch('source2', text_format, 'source2'),
-                    DatasetMatch('source3', text_format, 'source3')])
-        # TODO: Should test out other file formats as well.
-        source_files = [study.input(n)
-                        for n in ('source1', 'source2', 'source3')]
-        inputnode = pe.Node(
-            IdentityInterface(['subject_id', 'visit_id']), 'inputnode')
-        inputnode.inputs.subject_id = self.SUBJECT
-        inputnode.inputs.visit_id = self.VISIT
-        source = self.repository.source(source_files)
-        # Test subject sink
-        subject_sink_files = [
-            study.spec('subject_sink')]
-        subject_sink = self.repository.sink(subject_sink_files,
-                                         frequency='per_subject',
-                                         study_name=self.SUMMARY_STUDY_NAME)
-        subject_sink.inputs.name = 'subject_summary'
-        subject_sink.inputs.desc = (
-            "Tests the sinking of subject-wide datasets")
-        # Test visit sink
-        visit_sink_files = [study.spec('visit_sink')]
-        visit_sink = self.repository.sink(visit_sink_files,
-                                       frequency='per_visit',
-                                       study_name=self.SUMMARY_STUDY_NAME)
-        visit_sink.inputs.name = 'visit_summary'
-        visit_sink.inputs.desc = (
-            "Tests the sinking of visit-wide datasets")
-        # Test project sink
-        project_sink_files = [
-            study.spec('project_sink')]
-        project_sink = self.repository.sink(project_sink_files,
-                                         frequency='per_project',
-                                         study_name=self.SUMMARY_STUDY_NAME)
-
-        project_sink.inputs.name = 'project_summary'
-        project_sink.inputs.desc = (
-            "Tests the sinking of project-wide datasets")
-        # Create workflow connecting them together
-        workflow = pe.Workflow('summary_unittest', base_dir=self.work_dir)
-        workflow.add_nodes((source, subject_sink, visit_sink,
-                            project_sink))
-        workflow.connect(inputnode, 'subject_id', source, 'subject_id')
-        workflow.connect(inputnode, 'visit_id', source, 'visit_id')
-        workflow.connect(inputnode, 'subject_id', subject_sink, 'subject_id')
-        workflow.connect(inputnode, 'visit_id', visit_sink, 'visit_id')
-        workflow.connect(
-            source, 'source1' + PATH_SUFFIX,
-            subject_sink, 'subject_sink' + PATH_SUFFIX)
-        workflow.connect(
-            source, 'source2' + PATH_SUFFIX,
-            visit_sink, 'visit_sink' + PATH_SUFFIX)
-        workflow.connect(
-            source, 'source3' + PATH_SUFFIX,
-            project_sink, 'project_sink' + PATH_SUFFIX)
-        workflow.run()
-        # Check local summary directories were created properly
-        subject_dir = self.get_session_dir(frequency='per_subject')
-        self.assertEqual(sorted(os.listdir(subject_dir)),
-                         [self.SUMMARY_STUDY_NAME + '_subject_sink.txt'])
-        visit_dir = self.get_session_dir(frequency='per_visit')
-        self.assertEqual(sorted(os.listdir(visit_dir)),
-                         [self.SUMMARY_STUDY_NAME + '_visit_sink.txt'])
-        project_dir = self.get_session_dir(frequency='per_project')
-        self.assertEqual(sorted(os.listdir(project_dir)),
-                         [self.SUMMARY_STUDY_NAME + '_project_sink.txt'])
-        # Reload the data from the summary directories
-        reloadinputnode = pe.Node(IdentityInterface(['subject_id',
-                                                     'visit_id']),
-                                  'reload_inputnode')
-        reloadinputnode.inputs.subject_id = self.SUBJECT
-        reloadinputnode.inputs.visit_id = self.VISIT
-        reloadsource = self.repository.source(
-            (source_files + subject_sink_files + visit_sink_files +
-             project_sink_files),
-            name='reload_source',
-            study_name=self.SUMMARY_STUDY_NAME)
-        reloadsink = self.repository.sink(
-            [study.spec(n)
-             for n in ('resink1', 'resink2', 'resink3')],
-            study_name=self.SUMMARY_STUDY_NAME)
-        reloadsink.inputs.name = 'reload_summary'
-        reloadsink.inputs.desc = (
-            "Tests the reloading of subject and project summary datasets")
-        reloadworkflow = pe.Workflow('reload_summary_unittest',
-                                     base_dir=self.work_dir)
-        reloadworkflow.connect(reloadinputnode, 'subject_id',
-                               reloadsource, 'subject_id')
-        reloadworkflow.connect(reloadinputnode, 'visit_id',
-                               reloadsource, 'visit_id')
-        reloadworkflow.connect(reloadinputnode, 'subject_id',
-                               reloadsink, 'subject_id')
-        reloadworkflow.connect(reloadinputnode, 'visit_id',
-                               reloadsink, 'visit_id')
-        reloadworkflow.connect(reloadsource,
-                               'subject_sink' + PATH_SUFFIX,
-                               reloadsink,
-                               'resink1' + PATH_SUFFIX)
-        reloadworkflow.connect(reloadsource,
-                               'visit_sink' + PATH_SUFFIX,
-                               reloadsink,
-                               'resink2' + PATH_SUFFIX)
-        reloadworkflow.connect(reloadsource,
-                               'project_sink' + PATH_SUFFIX,
-                               reloadsink,
-                               'resink3' + PATH_SUFFIX)
-        reloadworkflow.run()
-        outputs = [
-            f for f in sorted(os.listdir(self.session_dir))
-            if f != FIELDS_FNAME]
-        self.assertEqual(outputs,
-                         [self.SUMMARY_STUDY_NAME + '_resink1.txt',
-                          self.SUMMARY_STUDY_NAME + '_resink2.txt',
-                          self.SUMMARY_STUDY_NAME + '_resink3.txt',
-                          'source1.txt', 'source2.txt',
-                          'source3.txt', 'source4.txt'])
-
-
 class TestProjectInfo(BaseMultiSubjectTestCase):
     """
     This unittest tests out that extracting the existing scans and
